+++ docs/changes.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Version 0.6.1
+-------------
+
+To be released.
+
+
 Version 0.6.0
 -------------